[delete] hard coded dummy band detail data
