+++ Reactify/ClientApp/src/components/bandDetail/BandDetail.css
@@ -1,0 +1,19 @@
+﻿.band-details {
+    margin-top: 50px;
+}
+
+.band-banner {
+;
+}
+
+.band-info {
+
+}
+
+.biography {
+
+}
+.band-empty {
+    margin-top: 80px;
+    text-align: center;
+}
