+++ Reactify/ClientApp/src/components/bandDetail/BandDetail.js
@@ -5,12 +5,12 @@ import Band from './Band.js';
 
 const BandDetail = () => {
     const [searchedBand, setSearchedBand] = useState({
-        "name": "Linkin Park",
-        "image": "https://images.unsplash.com/photo-1528645602411-bbeb0d69a6de?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=750&q=80",
-        "genre": "Alternative Rock",
-        "country": "California, USA",
-        "website": "www.linkinpark.com",
-        "biography": "Linkin Park is an American rock band from Agoura Hills, California, formed in 1996.",
+        "name": "",
+        "image": "",
+        "genre": "",
+        "country": "",
+        "website": "",
+        "biography": "",
     });
 
     const handleSubmit = () => {
